Collision of wild and domestic probably solved!

+++ src/Model/Constants.java
@@ -11,7 +11,7 @@ public class Constants
     public static final double ANIMAL_DISPLACEMENT_Y = Screen.getPrimary().getVisualBounds().getHeight()/ 3.75;
     public static final double ANIMAL_DISPLACEMENT_X = Screen.getPrimary().getVisualBounds().getWidth() / 4;
     public static final int ANIMAL_SIZE = 90;
-    public static final double scale = 3;
+    public static final int scale = 3;
     public static final int WELL_CAPACITY = 7;
     public static final int WAREHOUSE_VOLUME = 100;
     public static final int WELL_BASE_MONEY = 20;
